Create AssetGuardian.Registry
To divide AssetGuardian.Guardian's responsibility.
