+++ VitDeck/Assets/VitDeck/AssetGuardian/Registry.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 469367ec720a31844b03fa249685dc53
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
